+++ mountwizzard/mountwizzard.py
@@ -788,7 +788,6 @@ class MountWizzardApp(MwWidget):
                 self.coordinatePopup.ui.modellingLog.setText(self.coordinatePopup.ui.modellingLog.toPlainText() + text)     # otherwise add text at the end
             self.coordinatePopup.ui.modellingLog.moveCursor(QTextCursor.End)                                                # and move cursor up
             self.modelLogQueue.task_done()
-        app.instance().processEvents()
         # noinspection PyCallByClass,PyTypeChecker
         QTimer.singleShot(200, self.mainLoop)                                                                               # 200ms repeat time cyclic