+++ mountwizzard/support/model_thread.py
@@ -339,7 +339,7 @@ class Model(QtCore.QThread):
             ra = raCopy - float(i) * 6 / 20                                                                                 # 6 hours of track test
             az, alt = self.mount.transformNovas(ra, decCopy, 1)                                                             # transform to az alt
             if alt > 0:                                                                                                     # we only take point alt > 0
-                value.append((az, alt))                                                                                     # add point to list
+                value.append((int(az), int(alt)))                                                                                     # add point to list
         return value
 
     def generateDensePoints(self):                                                                                          # generate pointcloud in greater circles of sky
@@ -674,12 +674,8 @@ class Model(QtCore.QThread):
             return False, mes, modelData
         while True:                                                                                                         # retrieving solving data in loop
             suc, mes, ra_sol, dec_sol, scale, angle, timeTS = self.SGPro.SgGetSolvedImageData(guid)                         # retrieving the data from solver
-            modelData['dec_sol'] = float(dec_sol)                                                                           # convert to float
-            modelData['ra_sol'] = float(ra_sol)
-            modelData['scale'] = float(scale)
-            modelData['angle'] = float(angle)
-            modelData['timeTS'] = float(timeTS)
             mes = mes.strip('\n')                                                                                           # sometimes there are heading \n in message
+            print(mes)
             if mes[:7] in ['Matched', 'Solve t', 'Valid s']:                                                                # if there is success, we can move on
                 self.logger.debug('solveImage solv-> modelData {0}'.format(modelData))
                 solved = True
@@ -692,6 +688,11 @@ class Model(QtCore.QThread):
                     time.sleep(5)                                                                                           # therefore slow cycle
                 else:                                                                                                       # local solver takes 1-2 s
                     time.sleep(.25)                                                                                         # therefore quicker cycle
+        modelData['dec_sol'] = float(dec_sol)                                                                           # convert to float
+        modelData['ra_sol'] = float(ra_sol)
+        modelData['scale'] = float(scale)
+        modelData['angle'] = float(angle)
+        modelData['timeTS'] = float(timeTS)
         self.logger.debug('solveImage     -> suc:{0} mes:{1}'.format(suc, mes))                                             # debug output
         if solved:
             ra, dec = self.mount.transformNovas(modelData['ra_sol'], modelData['dec_sol'], 3)